% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admiralext-package.R
\docType{package}
\name{cicdworkshop-package}
\alias{cicdworkshop}
\alias{cicdworkshop-package}
\title{cicdworkshop: Demo Package for R/Pharma CI/CD Workshop}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is intended to demo CI/CD pipelines on GitHub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/admiraltemplate}
  \item \url{https://github.com/pharmaverse/admiraltemplate}
}

}
\author{
\strong{Maintainer}: Open Source \email{admiralext@pharmaverse.com}

}
\concept{internal}
\keyword{internal}
