% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello Workshop Attendees}
\usage{
hello()
}
\value{
Happy Message
}
\description{
Welcomes the CI/CD workshop attendees
}
\details{
Greet workshop attendees
}
\examples{
hello()
}
\author{
Yourself
}
