% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cicdworkshop-package.R
\docType{package}
\name{cicdworkshop-package}
\alias{cicdworkshop}
\alias{cicdworkshop-package}
\title{cicdworkshop: Demo Package for R/Pharma CI/CD Workshop}
\description{
This package is intended to demo CI/CD pipelines on GitHub.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pharmaverse.github.io/cicdworkshop.rinpharma2022}
  \item \url{https://github.com/pharmaverse/cicdworkshop.rinpharma2022}
}

}
\author{
\strong{Maintainer}: Open Source \email{osauthor@noreply.pharmaverse.com}

}
\concept{internal}
\keyword{internal}
